refactor(player): extract self-contained game clusters to player-game/ ES modules (#1279 step 6/6)
Decompose the player-game.js monolith by moving its most self-contained,
state-local clusters into ES submodules under www/js/player-game/.
player-game.js stays the module's public face and re-exports every function
consumers import — player-core.js and player-reveal.js import surfaces are
byte-identical (zero diff).

Extracted (state proven local to each cluster):
- player-game/timer.js (164 lines) — startCountdown, stopCountdown,
  _ensureTimerFloatObserver. State: countdownInterval, _timerFloatObserver,
  _timerFloatObservedTarget.
- player-game/artist-challenge.js (262 lines) — renderArtistChallenge,
  handleArtistGuess, handleArtistGuessAck, disableAllArtistButtons,
  showArtistResult, resetArtistChallengeState, renderArtistReveal. State:
  artistChallengeComplete, pendingArtistGuess, winningArtist,
  ARTIST_DEBOUNCE_MS, lastArtistGuessTime.
- player-game/movie-challenge.js (247 lines) — renderMovieChallenge,
  handleMovieGuess, handleMovieGuessAck, disableAllMovieButtons,
  showMovieResult, resetMovieChallengeState, renderMovieReveal. State:
  movieChallengeComplete, pendingMovieGuess, MOVIE_DEBOUNCE_MS,
  lastMovieGuessTime.

Cross-cluster wiring: updateGameView() and resetSubmissionState() (still in
player-game.js) call renderArtistChallenge/resetArtistChallengeState and the
movie equivalents — wired via explicit ES imports. The resulting import cycle
is event/runtime-driven only (no top-level cross-module call), so it is safe.

Skipped:
- steal cluster — its flag hasStealAvailable is written from
  resetSubmissionState() and read/written alongside hasSubmitted (submission
  cluster that stays in player-game.js). An ES-exported var is read-only across
  module boundaries, so steal cannot write the shared flag from another module.
  Keeping it in place is the conservative choice (matches step 4 deferrals).

player-game.js: 2306 -> 1684 lines (-622).

Verification: npm test 227 passed (unchanged); npm run build:check drift-free
(10/10 bundles match); node --check clean on all new source files. Bundle
node-eval throws the pre-existing "window is not defined" (browser global at
import) — identical on the baseline bundle, not a TDZ/circular-init regression.
