+++ custom_components/beatify/www/js/player-game.js
@@ -15,164 +15,31 @@ import {
     triggerConfetti, stopConfetti
 } from './player-utils.js';
 
+// #1279 step 6/6: self-contained game clusters extracted to ./player-game/.
+// player-game.js stays the public face and re-exports their consumer-facing
+// functions so player-core.js / player-reveal.js import surfaces are unchanged.
+export { startCountdown, stopCountdown } from './player-game/timer.js';
+export {
+    renderArtistChallenge, handleArtistGuessAck,
+    resetArtistChallengeState, renderArtistReveal
+} from './player-game/artist-challenge.js';
+export {
+    renderMovieChallenge, handleMovieGuessAck,
+    resetMovieChallengeState, renderMovieReveal
+} from './player-game/movie-challenge.js';
+
+// Cross-cluster calls: updateGameView() and resetSubmissionState() (still in
+// this file) call into the extracted artist/movie clusters. These are runtime
+// (event-driven), not module-init, so the circular import is safe.
+import {
+    renderArtistChallenge, resetArtistChallengeState
+} from './player-game/artist-challenge.js';
+import {
+    renderMovieChallenge, resetMovieChallengeState
+} from './player-game/movie-challenge.js';
+
 var utils = window.BeatifyUtils || {};
 var debug = utils.debug || function() {};
-
-// ============================================
-// Countdown Timer (Story 4.2)
-// ============================================
-
-var countdownInterval = null;
-
-/**
- * Start countdown timer
- * @param {number} deadline - Server deadline timestamp in milliseconds
- */
-export function startCountdown(deadline) {
-    stopCountdown();
-
-    var timerElement = document.getElementById('timer');
-    if (!timerElement) return;
-
-    var timerNeon = document.getElementById('timer-neon');
-    // #817: floating mini-timer for when the main timer scrolls out of view.
-    var timerFloat = document.getElementById('timer-float');
-    var timerFloatNum = document.getElementById('timer-float-num');
-
-    timerElement.classList.remove('timer--warning', 'timer--critical');
-    if (timerNeon) timerNeon.classList.remove('timer-neon--warn');
-    if (timerFloat) timerFloat.classList.remove('timer-float--warn');
-
-    // #817: arm the IntersectionObserver once per countdown. Shows the
-    // floating mini-timer when the main neon timer is NOT in viewport
-    // (typical when user scrolls down to reach the Submit button) and
-    // hides it when scrolled back up. Tear down on stopCountdown.
-    _ensureTimerFloatObserver(timerNeon, timerFloat);
-
-    // Watchdog tick counter — counts updateCountdown ticks spent past the
-    // deadline so the round_timeout nudge can retry instead of firing once.
-    var timedOutTicks = 0;
-
-    function updateCountdown() {
-        var now = Date.now();
-        var remaining = Math.max(0, Math.ceil((deadline - now) / 1000));
-
-        timerElement.textContent = remaining;
-        if (timerFloatNum) timerFloatNum.textContent = remaining;
-
-        if (remaining <= 5) {
-            timerElement.classList.remove('timer--warning');
-            timerElement.classList.add('timer--critical');
-        } else if (remaining <= 10) {
-            timerElement.classList.remove('timer--critical');
-            timerElement.classList.add('timer--warning');
-        } else {
-            timerElement.classList.remove('timer--warning', 'timer--critical');
-        }
-
-        // Arcade timer neon ring + floating pill: pink by default, red + pulse at ≤10s
-        if (timerNeon) {
-            timerNeon.classList.toggle('timer-neon--warn', remaining <= 10);
-        }
-        if (timerFloat) {
-            timerFloat.classList.toggle('timer-float--warn', remaining <= 10);
-        }
-
-        // ARIA announcements at key moments (Story 9.7)
-        if (remaining === 10) {
-            timerElement.setAttribute('aria-label', '10 seconds remaining');
-        } else if (remaining === 5) {
-            timerElement.setAttribute('aria-label', '5 seconds!');
-        } else if (remaining === 0) {
-            timerElement.setAttribute('aria-label', 'Time is up!');
-        } else {
-            timerElement.setAttribute('aria-label', 'Time remaining: ' + remaining + ' seconds');
-        }
-
-        if (remaining <= 0) {
-            // Watchdog: the server's round timer is a single async task — if
-            // it dies the round freezes on PLAYING forever (cancelled on a
-            // pause and never restarted, lost to a resume/desync edge). Our
-            // countdown is independent, so once it passes zero we nudge the
-            // server to end the round. handle_round_timeout is idempotent and
-            // only acts once the deadline truly passed — so a single nudge can
-            // race (clock skew) or be dropped (socket mid-reconnect) with no
-            // recovery. Keep nudging every few seconds until the phase leaves
-            // PLAYING, which tears this countdown down (player-core.js). Do
-            // NOT stopCountdown() here — that would make this single-shot.
-            timedOutTicks += 1;
-            if (timedOutTicks === 1 || timedOutTicks % 3 === 0) {
-                if (state.ws && state.ws.readyState === WebSocket.OPEN) {
-                    state.ws.send(JSON.stringify({ type: 'round_timeout' }));
-                }
-            }
-        }
-    }
-
-    updateCountdown();
-    countdownInterval = setInterval(updateCountdown, 1000);
-}
-
-// #817: IntersectionObserver state, scoped to one observer reused across
-// rounds. Recreated lazily on first startCountdown call after stopCountdown
-// (e.g. between rounds the DOM nodes can disappear/reappear).
-var _timerFloatObserver = null;
-var _timerFloatObservedTarget = null;
-
-function _ensureTimerFloatObserver(timerNeon, timerFloat) {
-    if (!timerFloat || !timerNeon) return;
-    if (typeof IntersectionObserver === 'undefined') {
-        // Fallback for ancient browsers — just always show the float during
-        // PLAYING. stopCountdown will hide it.
-        timerFloat.classList.remove('hidden');
-        timerFloat.classList.add('timer-float--visible');
-        return;
-    }
-    // If we're already observing the same target, leave it alone.
-    if (_timerFloatObserver && _timerFloatObservedTarget === timerNeon) return;
-    if (_timerFloatObserver) _timerFloatObserver.disconnect();
-
-    _timerFloatObserver = new IntersectionObserver(function(entries) {
-        var entry = entries[0];
-        if (!entry) return;
-        // When the main timer is NOT visible, show the float; otherwise hide.
-        if (entry.isIntersecting) {
-            timerFloat.classList.add('hidden');
-            timerFloat.classList.remove('timer-float--visible');
-        } else {
-            timerFloat.classList.remove('hidden');
-            timerFloat.classList.add('timer-float--visible');
-        }
-    }, {
-        // Trigger as soon as any part of the main timer leaves the viewport.
-        threshold: 0.1,
-    });
-    _timerFloatObserver.observe(timerNeon);
-    _timerFloatObservedTarget = timerNeon;
-}
-
-/**
- * Stop countdown timer
- */
-export function stopCountdown() {
-    if (countdownInterval) {
-        clearInterval(countdownInterval);
-        countdownInterval = null;
-    }
-    // #817: hide the floating mini-timer between rounds. The main timer
-    // node may also be torn down by view transitions; safe to leave the
-    // observer in place — re-arming on the next startCountdown is cheap.
-    var timerFloat = document.getElementById('timer-float');
-    if (timerFloat) {
-        timerFloat.classList.add('hidden');
-        timerFloat.classList.remove('timer-float--visible', 'timer-float--warn');
-    }
-    if (_timerFloatObserver) {
-        _timerFloatObserver.disconnect();
-        _timerFloatObserver = null;
-        _timerFloatObservedTarget = null;
-    }
-}
 
 // ============================================
 // Game View (Story 4.2)
@@ -683,19 +550,6 @@ var hasSubmitted = false;
 var betActive = false;
 var hasStealAvailable = false;
 
-// Artist Challenge state (Story 20.5)
-var artistChallengeComplete = false;
-var pendingArtistGuess = null;
-var winningArtist = null;
-var ARTIST_DEBOUNCE_MS = 300;
-var lastArtistGuessTime = 0;
-
-// Movie Challenge state (Issue #28)
-var movieChallengeComplete = false;
-var pendingMovieGuess = null;
-var MOVIE_DEBOUNCE_MS = 500;
-var lastMovieGuessTime = 0;
-
 // Title & Artist Mode state (#1180)
 var titleArtistMode = false;
 var taInputWired = false;
@@ -1105,482 +959,6 @@ function resetTitleArtistState() {
     if (titleInput) { titleInput.value = ''; titleInput.disabled = false; }
     if (artistInput) { artistInput.value = ''; artistInput.disabled = false; }
     if (ackEl) { ackEl.textContent = ''; ackEl.classList.add('hidden'); }
-}
-
-// ============================================
-// Artist Challenge (Story 20.5)
-// ============================================
-
-/**
- * Render artist challenge UI
- * @param {Object} artistChallenge - Artist challenge data from server
- * @param {string} phase - Current game phase (PLAYING, REVEAL)
- */
-export function renderArtistChallenge(artistChallenge, phase) {
-    var container = document.getElementById('artist-challenge-container');
-    if (!container) return;
-
-    if (!artistChallenge || !artistChallenge.options) {
-        container.classList.add('hidden');
-        return;
-    }
-
-    container.classList.remove('hidden');
-
-    var optionsEl = document.getElementById('artist-options');
-    var resultEl = document.getElementById('artist-result');
-
-    var currentOptions = Array.from(optionsEl.querySelectorAll('.artist-option-btn'))
-        .map(function(btn) { return btn.dataset.artist; });
-    var newOptions = artistChallenge.options;
-
-    if (JSON.stringify(currentOptions) !== JSON.stringify(newOptions)) {
-        optionsEl.innerHTML = '';
-        newOptions.forEach(function(artist, index) {
-            var btn = document.createElement('button');
-            btn.className = 'artist-option-btn';
-            btn.dataset.artist = artist;
-            btn.dataset.index = index;
-            btn.textContent = artist;
-            btn.addEventListener('click', function() {
-                handleArtistGuess(artist);
-            });
-            optionsEl.appendChild(btn);
-        });
-    }
-
-    if (artistChallenge.winner) {
-        var buttons = optionsEl.querySelectorAll('.artist-option-btn');
-        buttons.forEach(function(btn) {
-            btn.classList.add('is-disabled');
-            btn.classList.remove('is-loading', 'is-wrong');
-
-            var correctArtist = artistChallenge.correct_artist || winningArtist;
-            if (correctArtist && btn.dataset.artist === correctArtist) {
-                btn.classList.add('is-winner');
-            }
-        });
-
-        if (artistChallenge.winner === state.playerName) {
-            var bonusPoints = artistChallenge.bonus_points || 5;
-            resultEl.textContent = (utils.t('artistChallenge.youGotIt') || 'You got it! +{points} points')
-                .replace('{points}', bonusPoints);
-            resultEl.className = 'artist-result is-winner';
-        } else {
-            var msg = (utils.t('artistChallenge.someoneBeatYou') || '{winner} got it first!')
-                .replace('{winner}', artistChallenge.winner);
-            resultEl.textContent = msg;
-            resultEl.className = 'artist-result is-late';
-        }
-        resultEl.classList.remove('hidden');
-        artistChallengeComplete = true;
-    } else if (!artistChallengeComplete) {
-        resultEl.classList.add('hidden');
-    }
-}
-
-/**
- * Handle artist guess button click
- * @param {string} artist - The artist name that was clicked
- */
-function handleArtistGuess(artist) {
-    var now = Date.now();
-    if (now - lastArtistGuessTime < ARTIST_DEBOUNCE_MS) return;
-    lastArtistGuessTime = now;
-
-    if (artistChallengeComplete) return;
-
-    var btn = document.querySelector('.artist-option-btn[data-artist="' + CSS.escape(artist) + '"]');
-    if (btn) {
-        btn.classList.add('is-loading');
-    }
-
-    pendingArtistGuess = artist;
-
-    try {
-        if (state.ws && state.ws.readyState === WebSocket.OPEN) {
-            state.ws.send(JSON.stringify({
-                type: 'artist_guess',
-                artist: artist
-            }));
-        }
-    } catch (e) {
-        console.error('Artist guess send failed:', e);
-        if (btn) {
-            btn.classList.remove('is-loading');
-        }
-        pendingArtistGuess = null;
-    }
-}
-
-/**
- * Handle artist_guess_ack from server (Story 20.3 protocol)
- * @param {Object} data - Ack response from server
- */
-export function handleArtistGuessAck(data) {
-    var btn = pendingArtistGuess
-        ? document.querySelector('.artist-option-btn[data-artist="' + CSS.escape(pendingArtistGuess) + '"]')
-        : null;
-
-    if (data.correct && data.first) {
-        winningArtist = pendingArtistGuess;
-        if (btn) {
-            btn.classList.remove('is-loading');
-            btn.classList.add('is-correct');
-            var badge = document.createElement('span');
-            badge.className = 'artist-points-badge';
-            badge.textContent = '+' + (data.bonus_points || 5);
-            btn.appendChild(badge);
-        }
-        disableAllArtistButtons();
-        var bonusText = (utils.t('artistChallenge.youGotIt') || 'You got it! +{points} points')
-            .replace('{points}', data.bonus_points || 5);
-        showArtistResult(bonusText, true);
-        artistChallengeComplete = true;
-
-    } else if (data.correct && !data.first) {
-        winningArtist = pendingArtistGuess;
-        if (btn) {
-            btn.classList.remove('is-loading');
-            btn.classList.add('is-correct');
-        }
-        disableAllArtistButtons();
-        var msg = (utils.t('artistChallenge.someoneBeatYou') || '{winner} got it first!')
-            .replace('{winner}', data.winner || 'Someone');
-        showArtistResult(msg, false);
-        artistChallengeComplete = true;
-
-    } else {
-        if (btn) {
-            btn.classList.remove('is-loading');
-            btn.classList.add('is-wrong', 'is-selected');
-        }
-        disableAllArtistButtons();
-        showArtistResult(utils.t('artistChallenge.wrongGuess') || 'Wrong guess!', false);
-        artistChallengeComplete = true;
-    }
-
-    pendingArtistGuess = null;
-}
-
-/**
- * Disable all artist option buttons
- */
-function disableAllArtistButtons() {
-    document.querySelectorAll('.artist-option-btn').forEach(function(btn) {
-        btn.classList.add('is-disabled');
-        btn.classList.remove('is-loading');
-    });
-}
-
-/**
- * Show artist challenge result message
- * @param {string} message - Result message to display
- * @param {boolean} isWinner - Whether current player won
- */
-function showArtistResult(message, isWinner) {
-    var resultEl = document.getElementById('artist-result');
-    if (resultEl) {
-        resultEl.textContent = message;
-        resultEl.className = 'artist-result ' + (isWinner ? 'is-winner' : 'is-late');
-        resultEl.classList.remove('hidden');
-    }
-}
-
-/**
- * Reset artist challenge state for new round
- */
-function resetArtistChallengeState() {
-    artistChallengeComplete = false;
-    pendingArtistGuess = null;
-    winningArtist = null;
-
-    var container = document.getElementById('artist-challenge-container');
-    if (container) container.classList.add('hidden');
-
-    var optionsEl = document.getElementById('artist-options');
-    if (optionsEl) optionsEl.innerHTML = '';
-
-    var resultEl = document.getElementById('artist-result');
-    if (resultEl) {
-        resultEl.classList.add('hidden');
-        resultEl.className = 'artist-result hidden';
-    }
-}
-
-/**
- * Render artist challenge reveal section (Story 20.6)
- * @param {Object} artistChallenge - Artist challenge data with correct_artist and winner
- * @param {string} currentPlayerName - Current player's name for comparison
- */
-export function renderArtistReveal(artistChallenge, currentPlayerName) {
-    var section = document.getElementById('artist-reveal-section');
-    if (!section) return;
-
-    if (!artistChallenge || !artistChallenge.correct_artist) {
-        section.classList.add('hidden');
-        return;
-    }
-
-    section.classList.remove('hidden');
-
-    var nameEl = document.getElementById('artist-reveal-name');
-    if (nameEl) {
-        nameEl.textContent = artistChallenge.correct_artist;
-    }
-
-    var winnerEl = document.getElementById('artist-reveal-winner');
-    if (winnerEl) {
-        if (artistChallenge.winner) {
-            winnerEl.classList.remove('hidden');
-            if (artistChallenge.winner === currentPlayerName) {
-                var bonusPoints = artistChallenge.bonus_points || 5;
-                winnerEl.textContent = (utils.t('artistChallenge.youGotIt') || 'You got it! +{points} points')
-                    .replace('{points}', bonusPoints);
-                winnerEl.className = 'artist-reveal-winner is-you';
-            } else {
-                var msg = (utils.t('artistChallenge.winnerWas') || '{winner} got it first!')
-                    .replace('{winner}', artistChallenge.winner);
-                winnerEl.textContent = msg;
-                winnerEl.className = 'artist-reveal-winner is-other';
-            }
-        } else {
-            winnerEl.textContent = utils.t('artistChallenge.noWinner') || 'No one guessed the artist';
-            winnerEl.className = 'artist-reveal-winner artist-reveal-no-winner';
-            winnerEl.classList.remove('hidden');
-        }
-    }
-}
-
-// ============================================
-// Movie Challenge (Issue #28)
-// ============================================
-
-/**
- * Render movie challenge UI
- * @param {Object} movieChallenge - Movie challenge data from server
- * @param {string} phase - Current game phase (PLAYING, REVEAL)
- */
-export function renderMovieChallenge(movieChallenge, phase) {
-    var container = document.getElementById('movie-challenge-container');
-    if (!container) return;
-
-    if (!movieChallenge || !movieChallenge.options) {
-        container.classList.add('hidden');
-        return;
-    }
-
-    container.classList.remove('hidden');
-
-    var optionsEl = document.getElementById('movie-options');
-    var resultEl = document.getElementById('movie-result');
-
-    var currentOptions = Array.from(optionsEl.querySelectorAll('.movie-option-btn'))
-        .map(function(btn) { return btn.dataset.movie; });
-    var newOptions = movieChallenge.options;
-
-    if (JSON.stringify(currentOptions) !== JSON.stringify(newOptions)) {
-        optionsEl.innerHTML = '';
-        newOptions.forEach(function(movie, index) {
-            var btn = document.createElement('button');
-            btn.className = 'movie-option-btn';
-            btn.dataset.movie = movie;
-            btn.dataset.index = index;
-            btn.textContent = movie;
-            btn.addEventListener('click', function() {
-                handleMovieGuess(movie);
-            });
-            optionsEl.appendChild(btn);
-        });
-    }
-
-    if (movieChallengeComplete) {
-        var buttons = optionsEl.querySelectorAll('.movie-option-btn');
-        buttons.forEach(function(btn) {
-            btn.classList.add('is-disabled');
-        });
-    }
-}
-
-/**
- * Handle movie guess button click
- * @param {string} movie - The movie name that was clicked
- */
-function handleMovieGuess(movie) {
-    var now = Date.now();
-    if (now - lastMovieGuessTime < MOVIE_DEBOUNCE_MS) return;
-    lastMovieGuessTime = now;
-
-    if (movieChallengeComplete) return;
-
-    var btn = document.querySelector('.movie-option-btn[data-movie="' + CSS.escape(movie) + '"]');
-    if (btn) {
-        btn.classList.add('is-loading');
-    }
-
-    pendingMovieGuess = movie;
-
-    try {
-        if (state.ws && state.ws.readyState === WebSocket.OPEN) {
-            state.ws.send(JSON.stringify({
-                type: 'movie_guess',
-                movie: movie
-            }));
-        }
-    } catch (e) {
-        console.error('Movie guess send failed:', e);
-        if (btn) {
-            btn.classList.remove('is-loading');
-        }
-        pendingMovieGuess = null;
-    }
-}
-
-/**
- * Handle movie_guess_ack from server (Issue #28)
- * @param {Object} data - Ack response from server
- */
-export function handleMovieGuessAck(data) {
-    var btn = pendingMovieGuess
-        ? document.querySelector('.movie-option-btn[data-movie="' + CSS.escape(pendingMovieGuess) + '"]')
-        : null;
-
-    if (data.already_guessed) {
-        if (btn) {
-            btn.classList.remove('is-loading');
-        }
-        showMovieResult(utils.t('movieChallenge.alreadyGuessed') || 'Already guessed!', false);
-        movieChallengeComplete = true;
-        disableAllMovieButtons();
-
-    } else if (data.correct) {
-        if (btn) {
-            btn.classList.remove('is-loading');
-            btn.classList.add('is-correct');
-            if (data.bonus > 0) {
-                var badge = document.createElement('span');
-                badge.className = 'movie-rank-badge';
-                badge.textContent = '+' + data.bonus;
-                btn.appendChild(badge);
-            }
-        }
-        disableAllMovieButtons();
-        var bonusText = (utils.t('movieChallenge.youGotIt') || 'Correct! #{rank} — +{bonus} points')
-            .replace('{rank}', data.rank || 1)
-            .replace('{bonus}', data.bonus || 0);
-        showMovieResult(bonusText, true);
-        movieChallengeComplete = true;
-
-    } else {
-        if (btn) {
-            btn.classList.remove('is-loading');
-            btn.classList.add('is-wrong', 'is-selected');
-        }
-        disableAllMovieButtons();
-        showMovieResult(utils.t('movieChallenge.wrongGuess') || 'Not quite...', false);
-        movieChallengeComplete = true;
-    }
-
-    pendingMovieGuess = null;
-}
-
-/**
- * Disable all movie option buttons
- */
-function disableAllMovieButtons() {
-    document.querySelectorAll('.movie-option-btn').forEach(function(btn) {
-        btn.classList.add('is-disabled');
-        btn.classList.remove('is-loading');
-    });
-}
-
-/**
- * Show movie challenge result message
- * @param {string} message - Result message to display
- * @param {boolean} isWinner - Whether current player guessed correctly
- */
-function showMovieResult(message, isWinner) {
-    var resultEl = document.getElementById('movie-result');
-    if (resultEl) {
-        resultEl.textContent = message;
-        resultEl.className = 'movie-result ' + (isWinner ? 'is-winner' : 'is-late');
-        resultEl.classList.remove('hidden');
-    }
-}
-
-/**
- * Reset movie challenge state for new round
- */
-function resetMovieChallengeState() {
-    movieChallengeComplete = false;
-    pendingMovieGuess = null;
-
-    var container = document.getElementById('movie-challenge-container');
-    if (container) container.classList.add('hidden');
-
-    var optionsEl = document.getElementById('movie-options');
-    if (optionsEl) optionsEl.innerHTML = '';
-
-    var resultEl = document.getElementById('movie-result');
-    if (resultEl) {
-        resultEl.classList.add('hidden');
-        resultEl.className = 'movie-result hidden';
-    }
-}
-
-/**
- * Render movie challenge reveal section (Issue #28)
- * @param {Object} movieChallenge - Movie challenge data with correct_movie and results
- * @param {string} currentPlayerName - Current player's name for comparison
- */
-export function renderMovieReveal(movieChallenge, currentPlayerName) {
-    var section = document.getElementById('movie-reveal-section');
-    if (!section) return;
-
-    if (!movieChallenge || !movieChallenge.correct_movie) {
-        section.classList.add('hidden');
-        return;
-    }
-
-    section.classList.remove('hidden');
-
-    var nameEl = document.getElementById('movie-reveal-name');
-    if (nameEl) {
-        nameEl.textContent = movieChallenge.correct_movie;
-    }
-
-    var winnersEl = document.getElementById('movie-reveal-winners');
-    if (winnersEl && movieChallenge.results) {
-        var winners = movieChallenge.results.winners || [];
-        if (winners.length > 0) {
-            winnersEl.innerHTML = '';
-            winnersEl.classList.remove('hidden');
-
-            var title = document.createElement('div');
-            title.className = 'movie-reveal-winners-title';
-            title.textContent = utils.t('movieChallenge.winnersTitle') || 'Movie Quiz Winners';
-            winnersEl.appendChild(title);
-
-            winners.forEach(function(winner) {
-                var entry = document.createElement('div');
-                entry.className = 'movie-reveal-winner-entry';
-                if (winner.name === currentPlayerName) {
-                    entry.classList.add('is-you');
-                } else {
-                    entry.classList.add('is-other');
-                }
-                entry.textContent = winner.name + ' — +' + winner.bonus + ' (' + winner.time + 's)';
-                winnersEl.appendChild(entry);
-            });
-        } else {
-            winnersEl.innerHTML = '';
-            winnersEl.classList.remove('hidden');
-            var noWinner = document.createElement('div');
-            noWinner.className = 'movie-reveal-no-winner';
-            noWinner.textContent = utils.t('movieChallenge.noWinner') || 'No one guessed the movie';
-            winnersEl.appendChild(noWinner);
-        }
-    }
 }
 
 // ============================================
